refactor(dvp-swap): extract shared cancel/reject refund-and-close path
